Update the names of all torch disaggregate models
