+++ nilmtk_contrib/torch/WindowGRU.py
