+++ nilmtk_contrib/torch/rnn.py
