+++ nilmtk_contrib/torch/rnn_attention.py
